Add some simple guards to the install script to not repeat work

+++ install.sh
@@ -2,16 +2,18 @@
 
 ##Download LLVM/CLang version 3.4 via tarball
 echo "CU2CL: Fetching LLVM Tarball"
-wget http://llvm.org/releases/3.4/llvm-3.4.src.tar.gz
+if [ ! -f llvm-3.4.src.tar.gz ]; then wget http://llvm.org/releases/3.4/llvm-3.4.src.tar.gz; fi
 echo "CU2CL: Fetching Clang Tarball"
-wget http://llvm.org/releases/3.4/clang-3.4.src.tar.gz
+if [ ! -f clang-3.4.src.tar.gz ]; then wget http://llvm.org/releases/3.4/clang-3.4.src.tar.gz; fi
 echo "CU2CL: Unzipping tarballs"
-tar -xzf llvm-3.4.src.tar.gz
+if [ ! -d llvm-3.4 ]; then tar -xzf llvm-3.4.src.tar.gz; fi
+if [ ! -d llvm-3.4/tools/clang ]; then
 tar -xzf clang-3.4.src.tar.gz
 mv clang-3.4 llvm-3.4/tools/clang
+fi
 
 ## Make Build Directories
-mkdir llvm-build
+if [ ! -d llvm-build ]; then mkdir llvm-build; fi
 
 ## Build LLVM and Clang -- REQUIRES CMAKE
 cd llvm-build
@@ -19,20 +21,20 @@ echo "CU2CL: Configuring LLVM/Clang build"
 cmake -DCMAKE_BUILD_TYPE=Release ../llvm-3.4
 ## adjust the argument to -j to suit available cores for speed
 echo "CU2CL: Building LLVM/Clang"
-make -j2
+make -j4
 
 ## Repair a few unlinked header files in this Clang Revision
 echo "CU2CL: Repairing Clang headers"
-ln -s "$(pwd)"/tools/clang/include/clang/Basic/AttrList.inc ../llvm-3.4/tools/clang/include/clang/Basic/AttrList.inc
-ln -s "$(pwd)"/tools/clang/include/clang/Basic/DiagnosticCommonKinds.inc ../llvm-3.4/tools/clang/include/clang/Basic/DiagnosticCommonKinds.inc
-ln -s "$(pwd)"/tools/clang/include/clang/AST/DeclNodes.inc ../llvm-3.4/tools/clang/include/clang/AST/DeclNodes.inc
-ln -s "$(pwd)"/tools/clang/include/clang/AST/StmtNodes.inc ../llvm-3.4/tools/clang/include/clang/AST/StmtNodes.inc
-ln -s "$(pwd)"/tools/clang/include/clang/AST/CommentCommandList.inc ../llvm-3.4/tools/clang/include/clang/AST/CommentCommandList.inc
-ln -s "$(pwd)"/tools/clang/include/clang/AST/Attrs.inc ../llvm-3.4/tools/clang/include/clang/AST/Attrs.inc
-ln -s "$(pwd)"/tools/clang/include/clang/Driver/Options.inc ../llvm-3.4/tools/clang/include/clang/Driver/Options.inc
+if [ ! -L ../llvm-3.4/tools/clang/include/clang/Basic/AttrList.inc ]; then ln -s "$(pwd)"/tools/clang/include/clang/Basic/AttrList.inc ../llvm-3.4/tools/clang/include/clang/Basic/AttrList.inc; fi
+if [ ! -L ../llvm-3.4/tools/clang/include/clang/Basic/DiagnosticCommonKinds.inc ]; then ln -s "$(pwd)"/tools/clang/include/clang/Basic/DiagnosticCommonKinds.inc ../llvm-3.4/tools/clang/include/clang/Basic/DiagnosticCommonKinds.inc; fi
+if [ ! -L ../llvm-3.4/tools/clang/include/clang/AST/DeclNodes.inc ]; then ln -s "$(pwd)"/tools/clang/include/clang/AST/DeclNodes.inc ../llvm-3.4/tools/clang/include/clang/AST/DeclNodes.inc; fi
+if [ ! -L ../llvm-3.4/tools/clang/include/clang/AST/StmtNodes.inc ]; then ln -s "$(pwd)"/tools/clang/include/clang/AST/StmtNodes.inc ../llvm-3.4/tools/clang/include/clang/AST/StmtNodes.inc; fi
+if [ ! -L ../llvm-3.4/tools/clang/include/clang/AST/CommentCommandList.inc ]; then ln -s "$(pwd)"/tools/clang/include/clang/AST/CommentCommandList.inc ../llvm-3.4/tools/clang/include/clang/AST/CommentCommandList.inc; fi
+if [ ! -L ../llvm-3.4/tools/clang/include/clang/AST/Attrs.inc ]; then ln -s "$(pwd)"/tools/clang/include/clang/AST/Attrs.inc ../llvm-3.4/tools/clang/include/clang/AST/Attrs.inc; fi
+if [ ! -L ../llvm-3.4/tools/clang/include/clang/Driver/Options.inc ]; then ln -s "$(pwd)"/tools/clang/include/clang/Driver/Options.inc ../llvm-3.4/tools/clang/include/clang/Driver/Options.inc; fi
 
 ## Make a directory for the cu2cl shared library.
-mkdir ../cu2cl-build
+if [ ! -d ../cu2cl-build ]; then mkdir ../cu2cl-build; fi
 cd ../cu2cl-build
 
 ## Use CMake to configure the build
